requirements.txt
faker
Django==1.7.7
django-bootstrap3==5.1.1
django-extensions==1.5.0
django-registration-redux==1.1
fake-factory==0.5.0
six==1.9.0
wsgiref==0.1.2
